requirements.txt
datasets
openai
torch
torchcodec
soundfile
librosa
numpy
noisereduce
python-dotenv
jiwer
openpyxl
streamlit
pandas